little update: creating a main file and generalisation of hough_transform
