+++ grid_processing/Makefile
@@ -5,13 +5,13 @@ CFLAGS = -Wall -Wextra -O3 `pkg-config --cflags sdl2 SDL2_image`
 LDFLAGS =
 LDLIBS = `pkg-config --libs sdl2 SDL2_image`
 
-SRC = hough_transform.c doubles_lists.c basic_utilities.c SDL_utilities.c
+SRC = main.c hough_transform.c doubles_lists.c basic_utilities.c SDL_utilities.c
 OBJ = ${SRC:.c=.o}
 DEP = ${SRC:.c=.d}
 
-hough_transform: ${OBJ}
+main: ${OBJ}
 
 clean:
 	${RM} ${OBJ}
 	${RM} ${DEP}
-	${RM} hough_transform
+	${RM} main
