+++ grid_processing/SDL_utilities.h
@@ -23,6 +23,6 @@ int pixel_to_bit_value(SDL_Surface* surf, Uint32 color);
 void get_array_of_pixels(SDL_Surface* surf, int pixel_bit_array[]);
 void event_loop();
 void show_surface_on_window(SDL_Surface* surf);
-void draw_lines_on_window(struct list* list_rho, struct list* list_theta, char *path, double surf_diag);
+void draw_lines_on_window(struct list* list_rho, struct list* list_theta, SDL_Surface* surf, double surf_diag);
 
 #endif
