+++ grid_processing/tests/black_grid.png
